Magics : grib_file_address_mode="byte_offset" not working for large files [MAGP-1357] Advanced wind plotting hangs when no values found above min speed[MAGP-1355]

+++ src/decoders/GribDecoder.h
@@ -275,7 +275,7 @@ protected:
     bool directionDone_;
 
     mutable grib_handle* current_handle_;
-    mutable int current_position_;
+    mutable unsigned long long current_position_;
     mutable string current_file_name_;
 
     string title_;
